Refactor reproduction semantics and shared utilities
