+++ src/xdrl/__init__.py
@@ -6,6 +6,7 @@ from xdrl.compatibility import (
     SUPPORTED_DEPENDENCIES,
     SUPPORTED_GIT_REVISIONS,
     SUPPORTED_PYTHON,
+    WORKFLOW_CONFORMANCE,
     CompatibilityBoundaryError,
     ConformanceCheck,
     ConformanceSuite,
@@ -13,34 +14,15 @@ from xdrl.compatibility import (
     PrivateAPIUsage,
     SupportLevel,
     VersionRequirement,
-    WORKFLOW_CONFORMANCE,
     installed_dependency_revisions,
     installed_dependency_versions,
     validate_runtime_compatibility,
 )
+from xdrl.digests import bytes_digest, module_digest, named_tensor_digest, repository_revision, tensor_digest
 from xdrl.interactions import (
-    AgentSelector,
-    CoalitionTerm,
     InteractionContract,
-    InteractionPhase,
-    InteractionTopology,
-    InternalComputationAxis,
-    InternalComputationSemantics,
-    InternalOccurrence,
-    InternalOccurrenceSelection,
     LifecycleEvent,
-    LifecycleEventType,
-    MultiAgentSemantics,
-    NamedReduction,
-    OccurrenceIdentityError,
-    RecurrentCollectorMode,
-    RecurrentSemantics,
-    RecurrentStateTransition,
     RuntimeInteractionContext,
-    SemanticTarget,
-    ValueDecompositionAxes,
-    ValueDecompositionKeys,
-    ValueDecompositionSemantics,
 )
 from xdrl.interventions import (
     Intervention,
@@ -63,6 +45,7 @@ from xdrl.observations import (
     TensorRetention,
 )
 from xdrl.provenance import (
+    WORKFLOW_PROVENANCE_SCHEMA_REVISION,
     ArtifactDigestAlgorithm,
     InputArtifactReference,
     InputArtifactRole,
@@ -70,12 +53,32 @@ from xdrl.provenance import (
     OutputArtifactDigest,
     OutputArtifactReference,
     OutputArtifactRole,
-    WORKFLOW_PROVENANCE_SCHEMA_REVISION,
     ProvenanceSchemaError,
     WorkflowCompatibilityEvidence,
     WorkflowExecutionEvidence,
     WorkflowPlanEvidence,
     WorkflowProvenance,
+)
+from xdrl.semantics import (
+    AgentSelector,
+    CoalitionTerm,
+    InteractionPhase,
+    InteractionTopology,
+    InternalComputationAxis,
+    InternalComputationSemantics,
+    InternalOccurrence,
+    InternalOccurrenceSelection,
+    LifecycleEventType,
+    MultiAgentSemantics,
+    NamedReduction,
+    OccurrenceIdentityError,
+    RecurrentCollectorMode,
+    RecurrentSemantics,
+    RecurrentStateTransition,
+    SemanticTarget,
+    ValueDecompositionAxes,
+    ValueDecompositionKeys,
+    ValueDecompositionSemantics,
 )
 from xdrl.tdhook import (
     OutputArtifactResolver,
@@ -108,15 +111,25 @@ except PackageNotFoundError:
     __version__ = "unknown version"
 
 __all__ = [
+    "PRIVATE_UPSTREAM_APIS",
+    "SUPPORTED_DEPENDENCIES",
+    "SUPPORTED_GIT_REVISIONS",
+    "SUPPORTED_PYTHON",
+    "SUPPORT_DEFINITIONS",
+    "WORKFLOW_CONFORMANCE",
+    "WORKFLOW_PROVENANCE_SCHEMA_REVISION",
     "AgentSelector",
-    "CoalitionTerm",
     "ArtifactDigestAlgorithm",
     "BatchSemantics",
+    "CoalitionTerm",
     "CompatibilityBoundaryError",
     "ConformanceCheck",
     "ConformanceSuite",
     "ContractModule",
+    "DimensionReduction",
     "GitRevisionRequirement",
+    "InputArtifactReference",
+    "InputArtifactRole",
     "InteractionContract",
     "InteractionPhase",
     "InteractionTopology",
@@ -131,68 +144,63 @@ __all__ = [
     "InterventionTarget",
     "InterventionTiming",
     "InterventionValidationError",
-    "InputArtifactReference",
-    "InputArtifactRole",
     "KeyPresence",
     "KeyRole",
     "KeySchema",
     "LifecycleEvent",
     "LifecycleEventType",
+    "ModelRole",
     "MultiAgentSemantics",
     "NamedReduction",
-    "OccurrenceIdentityError",
-    "ModelRole",
     "ObservationKind",
     "ObservationRecord",
     "ObservationTrace",
+    "OccurrenceIdentityError",
     "OutputArtifactDeclaration",
     "OutputArtifactDigest",
     "OutputArtifactReference",
     "OutputArtifactResolver",
     "OutputArtifactRole",
-    "PRIVATE_UPSTREAM_APIS",
-    "PrivateAPIUsage",
-    "ProvenanceSchemaError",
+    "PairedInterventionResult",
     "PairedWorkflowExecutionError",
     "PairedWorkflowValidationError",
-    "DimensionReduction",
-    "ReductionKind",
+    "PrivateAPIUsage",
+    "ProvenanceSchemaError",
     "RecurrentCollectorMode",
     "RecurrentSemantics",
     "RecurrentStateTransition",
+    "ReductionKind",
     "RetentionPolicy",
     "RuntimeInteractionContext",
     "SchemaValidationError",
     "SemanticTarget",
+    "SupportLevel",
+    "TDHookPairedWorkflowResult",
+    "TDHookWorkflowPairManifest",
+    "TDHookWorkflowResult",
+    "TDHookWorkflowRunner",
+    "TensorDictKey",
+    "TensorDictSchema",
+    "TensorRetention",
+    "TorchRLModule",
     "ValueDecompositionAxes",
     "ValueDecompositionKeys",
     "ValueDecompositionSemantics",
-    "SUPPORTED_DEPENDENCIES",
-    "SUPPORTED_GIT_REVISIONS",
-    "SUPPORTED_PYTHON",
-    "SUPPORT_DEFINITIONS",
-    "SupportLevel",
-    "PairedInterventionResult",
-    "TDHookWorkflowResult",
-    "TDHookPairedWorkflowResult",
-    "TDHookWorkflowPairManifest",
-    "TDHookWorkflowRunner",
-    "TensorDictSchema",
-    "TensorDictKey",
-    "TensorRetention",
-    "TorchRLModule",
     "VersionRequirement",
-    "WORKFLOW_PROVENANCE_SCHEMA_REVISION",
-    "WorkflowCompatibilityEvidence",
     "WorkflowArmReference",
+    "WorkflowCompatibilityEvidence",
     "WorkflowExecutionEvidence",
     "WorkflowPlanEvidence",
     "WorkflowProvenance",
     "WorkflowStepDifference",
-    "WORKFLOW_CONFORMANCE",
+    "bytes_digest",
     "installed_dependency_revisions",
     "installed_dependency_versions",
+    "module_digest",
+    "named_tensor_digest",
+    "repository_revision",
     "run_paired",
+    "tensor_digest",
     "validate_module",
     "validate_runtime_compatibility",
 ]
